+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Use Broccoli Funnel instead of WatchedDir
+
+### Fixed
+* Ensure tree is only pruned once
+
 ## [1.0.0] - 2016-10-06
 ### Added
 * Initial release